#6 refactor: api 로직은 상위 폴더에 위치
