+++ src/services/oauth.ts
@@ -1,7 +1,7 @@
 import axios from 'axios'
-import api from '../../../utils/api'
-import { stringify } from '../../../utils/query'
-import { MyProfile } from '../../types/api'
+import api from '../utils/api'
+import { stringify } from '../utils/query'
+import { MyProfileQuery } from './user'
 
 export const post_oauth_login = async (accessToken: string) => {
   try {
@@ -16,15 +16,12 @@ export const post_oauth_login = async (accessToken: string) => {
 export const get_oauth_logout = async () => {
   try {
     const result = await api.get('/api/v1/oauth/logout')
+    await MyProfileQuery.invalidate()
     return result
   } catch (e) {
     console.error(e)
     throw new Error('로그아웃이 실패하였습니다. post v1/oauth/logout')
   }
-}
-
-export const get_user_myProfile = async () => {
-  return await api.get<MyProfile>('/api/v1/user/myProfile')
 }
 
 export const post_kakao_oauth_token = async (code: string) => {
